move big map functionality to bigmap.js
